Aten::addForcefield() now sets the new forcefield as the default if no others are loaded. Fixes to parser - pattern& and bound& (PatternBound) variables were completely missing. Corrected some small bugs in parser. Reinstated DL_POLY FIELD export. Changed the way non-standard pointer arrays (i.e. lists) are handled by ReturnValue. Fixed 'readdoublearray' and 'readintegerarray' to handle new array stuff. Fixed deselection of integer ranges of atom IDs. Tweaked appearance of progress indicator in GUI. ReturnValue now stores copy of text data in a Dnchar, rather than just the const char* pointer. Fixed operation of nested Format::writeStrings() - bad use of static members.

+++ src/base/lineparser.cpp
@@ -608,7 +608,7 @@ int LineParser::getInteger(int nbytes)
 }
 
 // Read an array of integer values from an (unformatted) input file
-bool LineParser::getIntegerArray(double *array, int count)
+bool LineParser::getIntegerArray(int *array, int count)
 {
 	file_.read((char*) array, count*sizeof(int));
 	if (file_.eof())
